move UwaziRelationshipList to domain folder
